Move proposal_block_from_template() to zebra-rpc and remove eyre dependency
