+++ zebra-rpc/src/methods/get_block_template_rpcs/types/get_block_template/proposal.rs
@@ -52,8 +52,47 @@ impl From<ProposalResponse> for Response {
     }
 }
 
-impl From<GetBlockTemplate> for Response {
-    fn from(template: GetBlockTemplate) -> Self {
-        Self::TemplateMode(Box::new(template))
+/// Returns a block proposal generated from a [`GetBlockTemplate`] RPC response.
+pub fn proposal_block_from_template(
+    GetBlockTemplate {
+        version,
+        height,
+        previous_block_hash: GetBlockHash(previous_block_hash),
+        default_roots:
+            DefaultRoots {
+                merkle_root,
+                block_commitments_hash,
+                ..
+            },
+        bits: difficulty_threshold,
+        coinbase_txn,
+        transactions: tx_templates,
+        ..
+    }: GetBlockTemplate,
+) -> Result<Block, SerializationError> {
+    if Height(height) > Height::MAX {
+        Err(SerializationError::Parse(
+            "height field must be lower than Height::MAX",
+        ))?;
+    };
+
+    let mut transactions = vec![coinbase_txn.data.as_ref().zcash_deserialize_into()?];
+
+    for tx_template in tx_templates {
+        transactions.push(tx_template.data.as_ref().zcash_deserialize_into()?);
     }
+
+    Ok(Block {
+        header: Arc::new(block::Header {
+            version,
+            previous_block_hash,
+            merkle_root,
+            commitment_bytes: block_commitments_hash.into(),
+            time: Utc::now(),
+            difficulty_threshold,
+            nonce: [0; 32],
+            solution: Solution::default(),
+        }),
+        transactions,
+    })
 }
